Conflicts:
	TSimulator.cpp

+++ TRandomGenerator.cpp
@@ -89,52 +89,56 @@ double TRandomGenerator::ran3(){
 //--------------------------------------------------------
 //Binomial Distribution
 //--------------------------------------------------------
-double TRandomGenerator::getBiomialRand(double pp, int n){
- int j;
- static int nold=(-1);
- double am,em,g,angle,p,bnl,sq,t,y;
- static double pold=(-1.0),pc,plog,pclog,en,oldg;
+double TRandomGenerator::getBiomialRand(double pp, long n){
+	int j;
+	static int nold=(-1);
+	double am,em,g,angle,p,bnl,sq,t,y;
+	static double pold=(-1.0),pc,plog,pclog,en,oldg;
 
- p=(pp <= 0.5 ? pp : 1.0-pp);
- am=n*p;
- if (n < 25) {
-  bnl=0;
-  for (j=1;j<=n;j++)
-   if (ran3() < p) ++bnl;
- } else if (am < 1.0) {
-  g=exp(-am);
-  t=1.0;
-  for (j=0;j<=n;j++) {
-   t *= ran3();
-   if (t < g) break;
-  }
-  bnl=(j <= n ? j : n);
- } else {
-  if (n != nold) {
-   en=n;
-   oldg=gammaln(en+1.0);
-   nold=n;
-  } if (p != pold) {
-   pc=1.0-p;
-   plog=log(p);
-   pclog=log(pc);
-   pold=p;
-  }
-  sq=sqrt(2.0*am*pc);
-  do {
-   do {
-    angle=3.141592654*ran3();
-    y=tan(angle);
-    em=sq*y+am;
-   } while (em < 0.0 || em >= (en+1.0));
-   em=floor(em);
-   t=1.2*sq*(1.0+y*y)*exp(oldg-gammaln(em+1.0)
-    -gammaln(en-em+1.0)+em*plog+(en-em)*pclog);
-  } while (ran3() > t);
-  bnl=em;
- }
- if (p != pp) bnl=n-bnl;
- return bnl;
+	p=(pp <= 0.5 ? pp : 1.0-pp);
+	am=n*p;
+
+	if (n < 25){
+		bnl=0;
+		for (j=1;j<=n;j++)
+			if (ran3() < p) ++bnl;
+	} else if (am < 1.0) {
+		g=exp(-am);
+		t=1.0;
+		for (j=0;j<=n;j++) {
+			t *= ran3();
+			if (t < g) break;
+		}
+		bnl=(j <= n ? j : n);
+	} else {
+		if (n != nold) {
+			en=n;
+			oldg=gammaln(en+1.0);
+			nold=n;
+		}
+		if (p != pold) {
+			pc=1.0-p;
+			plog=log(p);
+			pclog=log(pc);
+			pold=p;
+		}
+		sq=sqrt(2.0*am*pc);
+
+		do {
+			do {
+				angle=3.141592654*ran3();
+				y=tan(angle);
+				em=sq*y+am;
+			} while (em < 0.0 || em >= (en+1.0));
+
+			em=floor(em);
+			t=1.2*sq*(1.0+y*y)*exp(oldg-gammaln(em+1.0)
+					-gammaln(en-em+1.0)+em*plog+(en-em)*pclog);
+		} while (ran3() > t);
+		bnl=em;
+	}
+	if (p != pp) bnl=n-bnl;
+	return bnl;
 }
 
 double TRandomGenerator::factorial(int n){
